backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
redis==5.0.1
kafka-python==2.0.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
faker==22.2.0
imbalanced-learn==0.11.0
